refactor(graph): extract store schema inspection
