+++ src/code_graph/store.ts
@@ -102,6 +102,21 @@ import {
   type StoredCodeGraph,
 } from './store_models.js';
 export * from './store_models.js';
+import {
+  MAXIMUM_CANONICAL_DATE_MILLISECONDS,
+  REMOVED_VIEWS_TABLE_SQL,
+  REMOVED_VIEW_CLEANUP_ADMISSION_CURSOR_KEY,
+  REMOVED_VIEW_CLEANUP_COLUMNS,
+  REMOVED_VIEW_CLEANUP_DUE_INDEX_SQL,
+  REMOVED_VIEW_CLEANUP_EPOCH_SEQUENCE_KEY,
+  REMOVED_VIEW_CLEANUP_TABLE_SQL,
+  REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS,
+} from './store_removed_view_schema_contracts.js';
+import {
+  removedViewAuthorityTableState,
+  removedViewCleanupRecordedRevision,
+  removedViewCleanupSchemaState,
+} from './store_removed_view_schema_inspection.js';
 import type {CodeGraphDatabaseSessionOptions, CodeGraphStoreShape} from './store_shape.js';
 export type {
   CodeGraphDatabaseSessionOptions,
@@ -115,21 +130,22 @@ import {
   LEGACY_SNAPSHOT_BUILD_OWNERS_CONTRACT,
   PERSISTENT_EXTENSION_TABLES,
   REMOVED_BETA30_INDEXES,
-  type PersistentExtensionTableContract,
-  type PersistentExtensionTableInspection,
-  type SqliteForeignKeyRow,
-  type SqliteIndexInfoRow,
-  type SqliteIndexListRow,
-  type SqliteTableColumnRow,
 } from './store_schema_contracts.js';
 export {
   CODE_GRAPH_PERSISTENT_EXTENSION_TABLE_NAMES,
   type CodeGraphPersistentSchemaMigrationPhase,
 } from './store_schema_contracts.js';
 import {
+  codeGraphPersistentExtensionSchemaCompatible,
+  inspectPersistentExtensionTables,
+  persistentExtensionTableInspection,
+  validatePersistentExtensionTables,
+} from './store_schema_inspection.js';
+export {codeGraphPersistentExtensionSchemaCompatible};
+import {normalizeSchemaDefinition} from './store_schema_normalization.js';
+import {
   REMOVED_VIEW_CLEANUP_CURRENT_MAXIMUM_METADATA_ROWS,
   REMOVED_VIEW_CLEANUP_LEGACY_MAXIMUM_METADATA_ROWS,
-  SCHEMA_METADATA_TABLE_SQL,
   assertCodeGraphRuntimeSchemaCompatible,
   inspectBoundedSchemaMetadataRowCount,
   inspectBoundedSchemaMetadataValue,
@@ -2100,696 +2116,6 @@ export class CodeGraphStore extends Context.Service<CodeGraphStore, CodeGraphSto
   );
 }
 
-const REMOVED_VIEW_CLEANUP_ADMISSION_CURSOR_KEY = 'removed_view_cleanup_admission_cursor';
-const REMOVED_VIEW_CLEANUP_EPOCH_SEQUENCE_KEY = 'removed_view_cleanup_epoch_sequence';
-const MAXIMUM_CANONICAL_DATE_MILLISECONDS = 253_402_300_799_999;
-
-const REMOVED_VIEWS_TABLE_SQL = `CREATE TABLE IF NOT EXISTS removed_views (
-  worktree_id TEXT PRIMARY KEY NOT NULL,
-  expected_snapshot_id TEXT NOT NULL,
-  removed_at TEXT NOT NULL
-) WITHOUT ROWID`;
-
-const REMOVED_VIEW_CLEANUP_TABLE_SQL = `CREATE TABLE IF NOT EXISTS removed_view_cleanup (
-  worktree_id TEXT NOT NULL CHECK (
-    typeof(worktree_id) = 'text' AND length(CAST(worktree_id AS BLOB)) = 64
-  ),
-  expected_snapshot_id TEXT NOT NULL CHECK (
-    typeof(expected_snapshot_id) = 'text'
-    AND length(CAST(expected_snapshot_id AS BLOB)) BETWEEN 45 AND 67
-  ),
-  removed_at TEXT NOT NULL CHECK (
-    typeof(removed_at) = 'text' AND length(CAST(removed_at AS BLOB)) = 24
-  ),
-  epoch INTEGER NOT NULL CHECK (
-    typeof(epoch) = 'integer' AND epoch BETWEEN 1 AND 9007199254740991
-  ),
-  repository_id TEXT CHECK (
-    repository_id IS NULL
-    OR (typeof(repository_id) = 'text' AND length(CAST(repository_id AS BLOB)) = 64)
-  ),
-  provenance_record_digest TEXT CHECK (
-    provenance_record_digest IS NULL
-    OR (
-      typeof(provenance_record_digest) = 'text'
-      AND length(CAST(provenance_record_digest AS BLOB)) = 64
-    )
-  ),
-  provenance_record_identity TEXT CHECK (
-    provenance_record_identity IS NULL
-    OR (
-      typeof(provenance_record_identity) = 'text'
-      AND length(CAST(provenance_record_identity AS BLOB)) = 64
-    )
-  ),
-  phase TEXT NOT NULL CHECK (
-    typeof(phase) = 'text'
-    AND length(CAST(phase AS BLOB)) <= 15
-    AND phase IN ('vector-pointers', 'build-status', 'provenance', 'complete')
-  ),
-  cursor_token TEXT CHECK (
-    cursor_token IS NULL
-    OR (typeof(cursor_token) = 'text' AND length(CAST(cursor_token AS BLOB)) BETWEEN 1 AND 512)
-  ),
-  revision INTEGER NOT NULL CHECK (
-    typeof(revision) = 'integer' AND revision BETWEEN 0 AND 9007199254740991
-  ),
-  attempts INTEGER NOT NULL CHECK (
-    typeof(attempts) = 'integer' AND attempts BETWEEN 0 AND 9007199254740991
-  ),
-  next_attempt_at INTEGER NOT NULL CHECK (
-    typeof(next_attempt_at) = 'integer' AND next_attempt_at BETWEEN 0 AND 253402300799999
-  ),
-  blocked_code TEXT CHECK (
-    blocked_code IS NULL
-    OR (
-      typeof(blocked_code) = 'text'
-      AND length(CAST(blocked_code AS BLOB)) BETWEEN 1 AND 32
-      AND blocked_code IN (
-        'busy', 'evidence-unavailable', 'invalid-sidecar',
-        'io-error', 'permission-denied', 'schema-incompatible'
-      )
-    )
-  ),
-  updated_at TEXT NOT NULL CHECK (
-    typeof(updated_at) = 'text' AND length(CAST(updated_at AS BLOB)) = 24
-  ),
-  PRIMARY KEY (worktree_id, expected_snapshot_id),
-  CHECK (phase <> 'complete' OR (cursor_token IS NULL AND blocked_code IS NULL)),
-  CHECK (
-    (repository_id IS NULL AND provenance_record_digest IS NULL AND provenance_record_identity IS NULL)
-    OR (
-      repository_id IS NOT NULL
-      AND provenance_record_digest IS NOT NULL
-      AND provenance_record_identity IS NOT NULL
-    )
-  )
-) WITHOUT ROWID`;
-
-const REMOVED_VIEW_CLEANUP_STORED_TABLE_SQL = REMOVED_VIEW_CLEANUP_TABLE_SQL.replace(
-  'CREATE TABLE IF NOT EXISTS',
-  'CREATE TABLE',
-);
-const REMOVED_VIEW_CLEANUP_STORED_TABLE_SQL_LITERAL = `'${REMOVED_VIEW_CLEANUP_STORED_TABLE_SQL.replaceAll("'", "''")}'`;
-
-const REMOVED_VIEW_CLEANUP_DUE_INDEX_SQL = `CREATE INDEX IF NOT EXISTS removed_view_cleanup_due
-  ON removed_view_cleanup (next_attempt_at, worktree_id, expected_snapshot_id)
-  WHERE phase <> 'complete'`;
-
-const REMOVED_VIEW_CLEANUP_PRIMARY_KEY_TRIGGER_GUARD_SQL = `SELECT CASE
-      WHEN (
-        SELECT COUNT(*)
-        FROM (
-          SELECT name, type, tbl_name,
-                 CASE WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 8192 THEN sql ELSE NULL END AS sql
-          FROM sqlite_master
-          WHERE name = 'removed_view_cleanup' COLLATE NOCASE
-          LIMIT 2
-        )
-      ) <> 1
-      OR NOT EXISTS (
-        SELECT 1
-        FROM (
-          SELECT name, type, tbl_name,
-                 CASE WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 8192 THEN sql ELSE NULL END AS sql
-          FROM sqlite_master
-          WHERE name = 'removed_view_cleanup' COLLATE NOCASE
-          LIMIT 2
-        )
-        WHERE name = 'removed_view_cleanup'
-          AND type = 'table'
-          AND tbl_name = 'removed_view_cleanup'
-          AND sql = ${REMOVED_VIEW_CLEANUP_STORED_TABLE_SQL_LITERAL}
-      )
-      OR (
-        SELECT COUNT(*)
-        FROM (
-          SELECT seqno, cid, name, "desc", coll, "key"
-          FROM pragma_index_xinfo('sqlite_autoindex_removed_view_cleanup_1')
-          LIMIT 3
-        )
-        WHERE (
-          seqno = 0 AND cid = 0 AND name = 'worktree_id'
-            AND "desc" = 0 AND coll = 'BINARY' AND "key" = 1
-        ) OR (
-          seqno = 1 AND cid = 1 AND name = 'expected_snapshot_id'
-            AND "desc" = 0 AND coll = 'BINARY' AND "key" = 1
-        ) OR (
-          seqno = 2 AND cid = 2 AND name = 'removed_at'
-            AND "desc" = 0 AND coll = 'BINARY' AND "key" = 0
-        )
-      ) <> 3
-      THEN RAISE(ABORT, 'code graph removed view cleanup authority is incompatible')
-    END;`;
-
-const REMOVED_VIEW_CLEANUP_REVOKE_DELETE_TRIGGER_SQL = `CREATE TRIGGER IF NOT EXISTS removed_views_cleanup_revoke_delete
-  AFTER DELETE ON removed_views
-  BEGIN
-    ${REMOVED_VIEW_CLEANUP_PRIMARY_KEY_TRIGGER_GUARD_SQL}
-    DELETE FROM removed_view_cleanup
-    WHERE worktree_id = OLD.worktree_id
-      AND expected_snapshot_id = OLD.expected_snapshot_id
-      AND removed_at = OLD.removed_at;
-  END`;
-
-const SNAPSHOT_LEASE_BATON_INDEX_TRIGGER_GUARD_SQL = `SELECT CASE
-      WHEN (
-        SELECT COUNT(*)
-        FROM (
-          SELECT name, type, tbl_name,
-                 CASE WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 1024 THEN sql ELSE NULL END AS sql
-          FROM sqlite_master
-          WHERE name = 'snapshot_leases_snapshot_expiry' COLLATE NOCASE
-          LIMIT 2
-        )
-      ) <> 1
-      OR NOT EXISTS (
-        SELECT 1
-        FROM (
-          SELECT name, type, tbl_name,
-                 CASE WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 1024 THEN sql ELSE NULL END AS sql
-          FROM sqlite_master
-          WHERE name = 'snapshot_leases_snapshot_expiry' COLLATE NOCASE
-          LIMIT 2
-        )
-        WHERE name = 'snapshot_leases_snapshot_expiry'
-          AND type = 'index'
-          AND tbl_name = 'snapshot_leases'
-          AND typeof(sql) = 'text'
-          AND length(CAST(sql AS BLOB)) <= 1024
-          AND sql = 'CREATE INDEX snapshot_leases_snapshot_expiry ON snapshot_leases(snapshot_id, expires_at)'
-      )
-      OR (
-        SELECT COUNT(*)
-        FROM (
-          SELECT seqno, cid, name, "desc", coll, "key"
-          FROM pragma_index_xinfo('snapshot_leases_snapshot_expiry')
-          LIMIT 4
-        )
-      ) <> 3
-      OR (
-        SELECT COUNT(*)
-        FROM (
-          SELECT seqno, cid, name, "desc", coll, "key"
-          FROM pragma_index_xinfo('snapshot_leases_snapshot_expiry')
-          LIMIT 4
-        )
-        WHERE (
-          seqno = 0 AND name = 'snapshot_id' AND "desc" = 0 AND coll = 'BINARY' AND "key" = 1
-        ) OR (
-          seqno = 1 AND name = 'expires_at' AND "desc" = 0 AND coll = 'BINARY' AND "key" = 1
-        ) OR (
-          seqno = 2 AND cid = -1 AND name IS NULL AND "desc" = 0 AND coll = 'BINARY' AND "key" = 0
-        )
-      ) <> 3
-      OR EXISTS (
-        SELECT 1
-        FROM (
-          SELECT
-            typeof(token) AS token_type,
-            length(CAST(token AS BLOB)) AS token_bytes,
-            typeof(snapshot_id) AS snapshot_type,
-            length(CAST(snapshot_id AS BLOB)) AS snapshot_bytes,
-            CASE
-              WHEN typeof(expires_at) = 'integer' AND expires_at BETWEEN 0 AND 253402300799999
-              THEN 1 ELSE 0
-            END AS expires_valid,
-            CASE
-              WHEN typeof(retire_when_inactive) = 'integer' AND retire_when_inactive IN (0, 1)
-              THEN 1 ELSE 0
-            END AS retire_valid
-          FROM snapshot_leases INDEXED BY snapshot_leases_snapshot_expiry
-          WHERE snapshot_id = NEW.expected_snapshot_id
-            AND expires_at > CAST(strftime('%s', 'now') AS INTEGER) * 1000
-          ORDER BY expires_at
-          LIMIT 1
-        )
-        WHERE token_type <> 'text'
-          OR token_bytes NOT BETWEEN 1 AND 1024
-          OR snapshot_type <> 'text'
-          OR snapshot_bytes NOT BETWEEN 1 AND 1024
-          OR expires_valid <> 1
-          OR retire_valid <> 1
-      )
-      THEN RAISE(ABORT, 'code graph snapshot lease baton index is incompatible')
-    END;`
-  .replace(/\s+/gu, ' ')
-  .trim();
-
-const REMOVED_VIEW_CLEANUP_REVOKE_INSERT_TRIGGER_SQL = `CREATE TRIGGER IF NOT EXISTS removed_views_cleanup_revoke_insert
-  AFTER INSERT ON removed_views
-  BEGIN
-    ${REMOVED_VIEW_CLEANUP_PRIMARY_KEY_TRIGGER_GUARD_SQL}
-    ${SNAPSHOT_LEASE_BATON_INDEX_TRIGGER_GUARD_SQL}
-    DELETE FROM removed_view_cleanup
-    WHERE worktree_id = NEW.worktree_id
-      AND expected_snapshot_id = NEW.expected_snapshot_id;
-    UPDATE snapshot_leases
-    SET retire_when_inactive = 1
-    WHERE rowid = (
-      SELECT rowid
-      FROM snapshot_leases INDEXED BY snapshot_leases_snapshot_expiry
-      WHERE snapshot_id = NEW.expected_snapshot_id
-        AND expires_at > CAST(strftime('%s', 'now') AS INTEGER) * 1000
-      ORDER BY expires_at
-      LIMIT 1
-    );
-  END`;
-
-const REMOVED_VIEW_CLEANUP_REVOKE_UPDATE_TRIGGER_SQL = `CREATE TRIGGER IF NOT EXISTS removed_views_cleanup_revoke_update
-  AFTER UPDATE OF worktree_id, expected_snapshot_id, removed_at ON removed_views
-  WHEN OLD.worktree_id <> NEW.worktree_id
-    OR OLD.expected_snapshot_id <> NEW.expected_snapshot_id
-    OR OLD.removed_at <> NEW.removed_at
-  BEGIN
-    ${REMOVED_VIEW_CLEANUP_PRIMARY_KEY_TRIGGER_GUARD_SQL}
-    ${SNAPSHOT_LEASE_BATON_INDEX_TRIGGER_GUARD_SQL}
-    DELETE FROM removed_view_cleanup
-    WHERE worktree_id = OLD.worktree_id
-      AND expected_snapshot_id = OLD.expected_snapshot_id
-      AND removed_at = OLD.removed_at
-      AND phase <> 'complete';
-    DELETE FROM removed_view_cleanup
-    WHERE worktree_id = NEW.worktree_id
-      AND expected_snapshot_id = NEW.expected_snapshot_id;
-    UPDATE snapshot_leases
-    SET retire_when_inactive = 1
-    WHERE rowid = (
-      SELECT rowid
-      FROM snapshot_leases INDEXED BY snapshot_leases_snapshot_expiry
-      WHERE snapshot_id = NEW.expected_snapshot_id
-        AND expires_at > CAST(strftime('%s', 'now') AS INTEGER) * 1000
-      ORDER BY expires_at
-      LIMIT 1
-    );
-  END`;
-
-const REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS = [
-  {name: 'removed_views_cleanup_revoke_delete', sql: REMOVED_VIEW_CLEANUP_REVOKE_DELETE_TRIGGER_SQL},
-  {name: 'removed_views_cleanup_revoke_insert', sql: REMOVED_VIEW_CLEANUP_REVOKE_INSERT_TRIGGER_SQL},
-  {name: 'removed_views_cleanup_revoke_update', sql: REMOVED_VIEW_CLEANUP_REVOKE_UPDATE_TRIGGER_SQL},
-] as const;
-
-const REMOVED_VIEW_CLEANUP_COLUMNS = [
-  {name: 'worktree_id', notNull: true, primaryKeyPosition: 1, type: 'TEXT'},
-  {name: 'expected_snapshot_id', notNull: true, primaryKeyPosition: 2, type: 'TEXT'},
-  {name: 'removed_at', notNull: true, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'epoch', notNull: true, primaryKeyPosition: 0, type: 'INTEGER'},
-  {name: 'repository_id', notNull: false, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'provenance_record_digest', notNull: false, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'provenance_record_identity', notNull: false, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'phase', notNull: true, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'cursor_token', notNull: false, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'revision', notNull: true, primaryKeyPosition: 0, type: 'INTEGER'},
-  {name: 'attempts', notNull: true, primaryKeyPosition: 0, type: 'INTEGER'},
-  {name: 'next_attempt_at', notNull: true, primaryKeyPosition: 0, type: 'INTEGER'},
-  {name: 'blocked_code', notNull: false, primaryKeyPosition: 0, type: 'TEXT'},
-  {name: 'updated_at', notNull: true, primaryKeyPosition: 0, type: 'TEXT'},
-] as const;
-
-const removedViewAuthorityTableState = Effect.fn('codeGraph.removedViewAuthorityTableState')(function* (
-  sql: SqlClient.SqlClient,
-) {
-  const objects = yield* sql.unsafe<{
-    readonly name: unknown;
-    readonly sql: unknown;
-    readonly sql_bytes: unknown;
-    readonly tbl_name: unknown;
-    readonly type: unknown;
-  }>(
-    `SELECT name, type, tbl_name,
-            CASE
-              WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 512 THEN sql
-              ELSE NULL
-            END AS sql,
-            length(CAST(sql AS BLOB)) AS sql_bytes
-     FROM sqlite_master
-     WHERE name = 'removed_views' COLLATE NOCASE
-     LIMIT 2`,
-  );
-  if (objects.length === 0) return 'absent' as const;
-  const object = objects[0];
-  if (
-    objects.length !== 1 ||
-    object?.name !== 'removed_views' ||
-    object.type !== 'table' ||
-    object.tbl_name !== 'removed_views' ||
-    typeof object.sql_bytes !== 'number' ||
-    !Number.isSafeInteger(object.sql_bytes) ||
-    object.sql_bytes > 512 ||
-    typeof object.sql !== 'string' ||
-    normalizeSchemaDefinition(object.sql) !== normalizeSchemaDefinition(REMOVED_VIEWS_TABLE_SQL)
-  ) {
-    return 'incompatible' as const;
-  }
-  const columns = yield* sql.unsafe<{
-    readonly dflt_value: unknown;
-    readonly hidden: unknown;
-    readonly name: unknown;
-    readonly notnull: unknown;
-    readonly pk: unknown;
-    readonly type: unknown;
-  }>(`SELECT * FROM pragma_table_xinfo('removed_views') LIMIT 4`);
-  const expected = [
-    {name: 'worktree_id', pk: 1},
-    {name: 'expected_snapshot_id', pk: 0},
-    {name: 'removed_at', pk: 0},
-  ] as const;
-  if (
-    columns.length !== expected.length ||
-    columns.some((column, index) => {
-      const contract = expected[index];
-      return (
-        contract === undefined ||
-        column.name !== contract.name ||
-        column.type !== 'TEXT' ||
-        column.notnull !== 1 ||
-        column.pk !== contract.pk ||
-        column.hidden !== 0 ||
-        column.dflt_value !== null
-      );
-    }) ||
-    (yield* sql.unsafe(`SELECT 1 FROM pragma_foreign_key_list('removed_views') LIMIT 1`)).length !== 0
-  ) {
-    return 'incompatible' as const;
-  }
-  const indexes = yield* sql.unsafe<{
-    readonly name: unknown;
-    readonly origin: unknown;
-    readonly partial: unknown;
-    readonly unique: unknown;
-  }>(`SELECT name, origin, partial, "unique" AS "unique" FROM pragma_index_list('removed_views') LIMIT 2`);
-  if (
-    indexes.length !== 1 ||
-    indexes[0]?.name !== 'sqlite_autoindex_removed_views_1' ||
-    indexes[0]?.origin !== 'pk' ||
-    indexes[0]?.unique !== 1 ||
-    indexes[0]?.partial !== 0
-  ) {
-    return 'incompatible' as const;
-  }
-  const primary = yield* sql.unsafe<{
-    readonly cid: unknown;
-    readonly coll: unknown;
-    readonly desc: unknown;
-    readonly key: unknown;
-    readonly name: unknown;
-    readonly seqno: unknown;
-  }>(`SELECT * FROM pragma_index_xinfo('sqlite_autoindex_removed_views_1') LIMIT 4`);
-  return primary.length === 3 &&
-    primary[0]?.seqno === 0 &&
-    primary[0]?.cid === 0 &&
-    primary[0]?.name === 'worktree_id' &&
-    primary[0]?.desc === 0 &&
-    primary[0]?.coll === 'BINARY' &&
-    primary[0]?.key === 1 &&
-    primary
-      .slice(1)
-      .every(
-        (column, index) =>
-          column.seqno === index + 1 &&
-          column.cid === index + 1 &&
-          column.name === expected[index + 1]?.name &&
-          column.desc === 0 &&
-          column.coll === 'BINARY' &&
-          column.key === 0,
-      )
-    ? ('compatible' as const)
-    : ('incompatible' as const);
-});
-
-const removedViewCleanupSchemaState = Effect.fn('codeGraph.removedViewCleanupSchemaState')(function* (
-  sql: SqlClient.SqlClient,
-) {
-  const objects = yield* sql.unsafe<{
-    readonly name: string;
-    readonly sql: string | null;
-    readonly sql_bytes: number | null;
-    readonly tbl_name: string;
-    readonly type: string;
-  }>(
-    `SELECT type, name, tbl_name,
-            CASE
-              WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 8192 THEN sql
-              ELSE NULL
-            END AS sql,
-            length(CAST(sql AS BLOB)) AS sql_bytes
-     FROM sqlite_master
-     WHERE lower(name) IN ('removed_view_cleanup', 'removed_view_cleanup_due')
-        OR (type = 'trigger' AND tbl_name = 'removed_views' COLLATE NOCASE)
-     ORDER BY name, type
-     LIMIT 6`,
-  );
-  const expectedNames = new Set([
-    'removed_view_cleanup',
-    'removed_view_cleanup_due',
-    ...REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS.map(trigger => trigger.name),
-  ]);
-  if (
-    objects.some(
-      object =>
-        object.name !== object.name.toLowerCase() ||
-        !expectedNames.has(object.name) ||
-        typeof object.sql_bytes !== 'number' ||
-        !Number.isSafeInteger(object.sql_bytes) ||
-        object.sql_bytes > 8192,
-    )
-  ) {
-    return 'incompatible' as const;
-  }
-  const tables = objects.filter(object => object.name === 'removed_view_cleanup');
-  const dueObjects = objects.filter(object => object.name === 'removed_view_cleanup_due');
-  const triggerObjects = objects.filter(object =>
-    REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS.some(trigger => trigger.name === object.name),
-  );
-  if (tables.length === 0) {
-    return dueObjects.length === 0 && triggerObjects.length === 0 ? ('absent' as const) : ('incompatible' as const);
-  }
-  if (
-    tables.length !== 1 ||
-    tables[0]?.type !== 'table' ||
-    tables[0]?.tbl_name !== 'removed_view_cleanup' ||
-    typeof tables[0]?.sql !== 'string' ||
-    dueObjects.some(object => object.type !== 'index' || object.tbl_name !== 'removed_view_cleanup') ||
-    triggerObjects.length !== REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS.length ||
-    REMOVED_VIEW_CLEANUP_TRIGGER_DEFINITIONS.some(expected => {
-      const observed = triggerObjects.find(object => object.name === expected.name);
-      return (
-        observed?.type !== 'trigger' ||
-        observed.tbl_name !== 'removed_views' ||
-        typeof observed.sql !== 'string' ||
-        normalizeSchemaDefinition(observed.sql) !== normalizeSchemaDefinition(expected.sql)
-      );
-    })
-  ) {
-    return 'incompatible' as const;
-  }
-  const columns = yield* sql.unsafe<{
-    readonly hidden: number;
-    readonly name: string;
-    readonly notnull: number;
-    readonly pk: number;
-    readonly type: string;
-  }>(`SELECT * FROM pragma_table_xinfo('removed_view_cleanup') LIMIT ${REMOVED_VIEW_CLEANUP_COLUMNS.length + 1}`);
-  const compatibleColumns =
-    columns.length === REMOVED_VIEW_CLEANUP_COLUMNS.length &&
-    columns.every((column, index) => {
-      const expected = REMOVED_VIEW_CLEANUP_COLUMNS[index];
-      return (
-        expected !== undefined &&
-        Number(column.hidden) === 0 &&
-        column.name === expected.name &&
-        column.type.toUpperCase() === expected.type &&
-        Number(column.notnull) === Number(expected.notNull) &&
-        Number(column.pk) === expected.primaryKeyPosition
-      );
-    });
-  const compatibleDefinition =
-    normalizeSchemaDefinition(tables[0].sql) === normalizeSchemaDefinition(REMOVED_VIEW_CLEANUP_TABLE_SQL);
-  const foreignKeys = yield* sql.unsafe(`SELECT 1 FROM pragma_foreign_key_list('removed_view_cleanup') LIMIT 1`);
-  const triggers = yield* sql.unsafe(
-    `SELECT name FROM sqlite_master
-     WHERE type = 'trigger' AND tbl_name = 'removed_view_cleanup' COLLATE NOCASE
-     LIMIT 1`,
-  );
-  if (!compatibleColumns || !compatibleDefinition || foreignKeys.length !== 0 || triggers.length !== 0) {
-    return 'incompatible' as const;
-  }
-
-  const indexes = yield* sql.unsafe<{
-    readonly name: string;
-    readonly origin: string;
-    readonly partial: number;
-    readonly unique: number;
-  }>(`SELECT name, origin, partial, "unique" AS "unique"
-      FROM pragma_index_list('removed_view_cleanup') LIMIT 3`);
-  const primary = indexes.find(index => index.origin === 'pk');
-  const due = indexes.find(index => index.name === 'removed_view_cleanup_due');
-  if (due === undefined) {
-    // Building this index over an existing queue would be unbounded startup
-    // work. Revision 8 creates the empty table and index atomically instead.
-    return 'incompatible' as const;
-  }
-  if (primary?.name !== 'sqlite_autoindex_removed_view_cleanup_1') return 'incompatible' as const;
-  const primaryColumns = yield* sql.unsafe<{
-    readonly coll: unknown;
-    readonly desc: unknown;
-    readonly key: unknown;
-    readonly name: unknown;
-    readonly seqno: unknown;
-  }>(
-    `SELECT seqno, name, desc, coll, key
-     FROM pragma_index_xinfo('sqlite_autoindex_removed_view_cleanup_1')
-     LIMIT ${REMOVED_VIEW_CLEANUP_COLUMNS.length + 1}`,
-  );
-  if (
-    primaryColumns.length !== REMOVED_VIEW_CLEANUP_COLUMNS.length ||
-    primaryColumns.some(
-      (column, index) =>
-        column.seqno !== index ||
-        column.name !== REMOVED_VIEW_CLEANUP_COLUMNS[index]?.name ||
-        column.desc !== 0 ||
-        column.coll !== 'BINARY' ||
-        column.key !== (index < 2 ? 1 : 0),
-    )
-  ) {
-    return 'incompatible' as const;
-  }
-  return dueObjects.length === 1 &&
-    indexes.length === 2 &&
-    primary !== undefined &&
-    Number(primary.unique) === 1 &&
-    Number(primary.partial) === 0 &&
-    Number(due.unique) === 0 &&
-    due.origin === 'c' &&
-    Number(due.partial) === 1 &&
-    normalizeSchemaDefinition(dueObjects[0]?.sql ?? '') ===
-      normalizeSchemaDefinition(REMOVED_VIEW_CLEANUP_DUE_INDEX_SQL)
-    ? ('compatible' as const)
-    : ('incompatible' as const);
-});
-
-const removedViewCleanupRecordedRevision = Effect.fn('codeGraph.removedViewCleanupRecordedRevision')(function* (
-  sql: SqlClient.SqlClient,
-) {
-  const metadataObjects = yield* sql.unsafe<{
-    readonly name: unknown;
-    readonly sql: unknown;
-    readonly sql_bytes: unknown;
-    readonly sql_type: unknown;
-    readonly tbl_name: unknown;
-    readonly type: unknown;
-  }>(
-    `SELECT name, type, tbl_name,
-            CASE
-              WHEN typeof(sql) = 'text' AND length(CAST(sql AS BLOB)) <= 256 THEN sql
-              ELSE NULL
-            END AS sql,
-            typeof(sql) AS sql_type,
-            length(CAST(sql AS BLOB)) AS sql_bytes
-     FROM sqlite_master
-     WHERE name = 'schema_metadata' COLLATE NOCASE
-     ORDER BY type
-     LIMIT 2`,
-  );
-  if (metadataObjects.length === 0) return {metadataPresent: false, state: 'missing'};
-  if (
-    metadataObjects.length !== 1 ||
-    metadataObjects[0]?.name !== 'schema_metadata' ||
-    metadataObjects[0]?.type !== 'table' ||
-    metadataObjects[0]?.tbl_name !== 'schema_metadata' ||
-    metadataObjects[0]?.sql_type !== 'text' ||
-    typeof metadataObjects[0]?.sql_bytes !== 'number' ||
-    !Number.isSafeInteger(metadataObjects[0].sql_bytes) ||
-    metadataObjects[0].sql_bytes > 256 ||
-    typeof metadataObjects[0]?.sql !== 'string' ||
-    normalizeSchemaDefinition(metadataObjects[0].sql) !== normalizeSchemaDefinition(SCHEMA_METADATA_TABLE_SQL)
-  ) {
-    return {state: 'invalid'};
-  }
-  const columns = yield* sql.unsafe<{
-    readonly dflt_value: unknown;
-    readonly hidden: unknown;
-    readonly name: unknown;
-    readonly notnull: unknown;
-    readonly pk: unknown;
-    readonly type: unknown;
-  }>(`SELECT * FROM pragma_table_xinfo('schema_metadata') LIMIT 3`);
-  if (
-    columns.length !== 2 ||
-    columns[0]?.name !== 'key' ||
-    columns[0]?.type !== 'TEXT' ||
-    Number(columns[0]?.notnull) !== 1 ||
-    Number(columns[0]?.pk) !== 1 ||
-    Number(columns[0]?.hidden) !== 0 ||
-    columns[0]?.dflt_value !== null ||
-    columns[1]?.name !== 'value' ||
-    columns[1]?.type !== 'TEXT' ||
-    Number(columns[1]?.notnull) !== 1 ||
-    Number(columns[1]?.pk) !== 0 ||
-    Number(columns[1]?.hidden) !== 0 ||
-    columns[1]?.dflt_value !== null
-  ) {
-    return {state: 'invalid'};
-  }
-  const indexes = yield* sql.unsafe<{
-    readonly name: unknown;
-    readonly origin: unknown;
-    readonly partial: unknown;
-    readonly unique: unknown;
-  }>(`SELECT name, origin, partial, "unique" AS "unique" FROM pragma_index_list('schema_metadata') LIMIT 2`);
-  if (
-    indexes.length !== 1 ||
-    indexes[0]?.name !== 'sqlite_autoindex_schema_metadata_1' ||
-    indexes[0]?.origin !== 'pk' ||
-    Number(indexes[0]?.unique) !== 1 ||
-    Number(indexes[0]?.partial) !== 0
-  ) {
-    return {state: 'invalid'};
-  }
-  const keyIndex = yield* sql.unsafe<{
-    readonly cid: unknown;
-    readonly coll: unknown;
-    readonly desc: unknown;
-    readonly key: unknown;
-    readonly name: unknown;
-    readonly seqno: unknown;
-  }>(`SELECT * FROM pragma_index_xinfo('sqlite_autoindex_schema_metadata_1') LIMIT 3`);
-  if (
-    keyIndex.length !== 2 ||
-    Number(keyIndex[0]?.seqno) !== 0 ||
-    Number(keyIndex[0]?.cid) !== 0 ||
-    keyIndex[0]?.name !== 'key' ||
-    Number(keyIndex[0]?.desc) !== 0 ||
-    keyIndex[0]?.coll !== 'BINARY' ||
-    Number(keyIndex[0]?.key) !== 1 ||
-    Number(keyIndex[1]?.seqno) !== 1 ||
-    Number(keyIndex[1]?.cid) !== -1 ||
-    keyIndex[1]?.name !== null ||
-    Number(keyIndex[1]?.desc) !== 0 ||
-    keyIndex[1]?.coll !== 'BINARY' ||
-    Number(keyIndex[1]?.key) !== 0
-  ) {
-    return {state: 'invalid'};
-  }
-  const foreignKeys = yield* sql.unsafe(`SELECT 1 FROM pragma_foreign_key_list('schema_metadata') LIMIT 1`);
-  const triggers = yield* sql.unsafe(
-    "SELECT name FROM sqlite_master WHERE type = 'trigger' AND tbl_name = 'schema_metadata' COLLATE NOCASE LIMIT 1",
-  );
-  if (foreignKeys.length !== 0 || triggers.length !== 0) return {state: 'invalid'};
-
-  const revision = yield* inspectBoundedSchemaMetadataValue(sql, 'persistent_extension_schema_revision', 16);
-  if (revision.state === 'missing') return {metadataPresent: true, state: 'missing'};
-  if (
-    revision.state === 'invalid' ||
-    !/^(?:0|[1-9][0-9]*)$/u.test(revision.value) ||
-    !Number.isSafeInteger(Number(revision.value))
-  ) {
-    return {state: 'invalid'};
-  }
-  return {state: 'recorded', value: Number(revision.value)};
-});
-
 const preflightRemovedViewCleanupSchema = Effect.fn('codeGraph.preflightRemovedViewCleanupSchema')(function* (
   sql: SqlClient.SqlClient,
 ) {
@@ -2974,118 +2300,6 @@ const ensureRemovedViewCleanupSchema = Effect.fn('codeGraph.ensureRemovedViewCle
     return yield* Effect.fail(new CodeGraphStoreError('Code graph removed view cleanup schema is incompatible.'));
   }
 });
-function persistentExtensionTableInspection(
-  sql: SqlClient.SqlClient,
-  contract: PersistentExtensionTableContract,
-): Effect.Effect<PersistentExtensionTableInspection, SqlError.SqlError> {
-  return Effect.gen(function* () {
-    const definitions = yield* sql<{readonly sql: string}>`
-      SELECT sql FROM sqlite_master WHERE type = 'table' AND name = ${contract.name} LIMIT 1
-    `;
-    const definition = definitions[0]?.sql;
-    if (definition === undefined) {
-      return {compatible: false, exists: false, group: contract.group, name: contract.name};
-    }
-    const columns = yield* sql.unsafe<SqliteTableColumnRow>(`PRAGMA table_info("${contract.name}")`);
-    const foreignKeys = yield* sql.unsafe<SqliteForeignKeyRow>(`PRAGMA foreign_key_list("${contract.name}")`);
-    const indexes =
-      contract.uniqueKeys === undefined
-        ? []
-        : yield* sql.unsafe<SqliteIndexListRow>(`PRAGMA index_list("${contract.name}")`);
-    const compatibleColumns =
-      columns.length === contract.columns.length &&
-      columns.every((column, index) => {
-        const expected = contract.columns[index];
-        return (
-          expected !== undefined &&
-          Number(column.cid) === index &&
-          column.name === expected.name &&
-          column.type.toUpperCase() === expected.type &&
-          Number(column.notnull) === Number(expected.notNull) &&
-          Number(column.pk) === expected.primaryKeyPosition &&
-          column.dflt_value == null
-        );
-      });
-    const expectedForeignKeys = contract.foreignKeys ?? [
-      {from: 'snapshot_id', onDelete: 'CASCADE', table: 'snapshots', to: 'id'},
-    ];
-    const actualForeignKeys = foreignKeys
-      .map(key => ({
-        from: key.from,
-        onDelete: key.on_delete.toUpperCase(),
-        table: key.table,
-        to: key.to,
-      }))
-      .sort(
-        (left, right) =>
-          compareCodeUnits(left.from, right.from) ||
-          compareCodeUnits(left.table, right.table) ||
-          compareCodeUnits(left.to, right.to),
-      );
-    const normalizedExpectedForeignKeys = [...expectedForeignKeys]
-      .map(key => ({...key, onDelete: key.onDelete.toUpperCase()}))
-      .sort(
-        (left, right) =>
-          compareCodeUnits(left.from, right.from) ||
-          compareCodeUnits(left.table, right.table) ||
-          compareCodeUnits(left.to, right.to),
-      );
-    const compatibleForeignKeys =
-      actualForeignKeys.length === normalizedExpectedForeignKeys.length &&
-      actualForeignKeys.every((key, index) => {
-        const expected = normalizedExpectedForeignKeys[index];
-        return (
-          expected !== undefined &&
-          key.from === expected.from &&
-          key.table === expected.table &&
-          key.to === expected.to &&
-          key.onDelete === expected.onDelete
-        );
-      });
-    const actualUniqueKeys: (readonly string[])[] = [];
-    for (const index of indexes) {
-      if (Number(index.unique) !== 1 || Number(index.partial) !== 0) continue;
-      const escapedIndexName = index.name.replaceAll('"', '""');
-      const indexedColumns = yield* sql.unsafe<SqliteIndexInfoRow>(`PRAGMA index_info("${escapedIndexName}")`);
-      actualUniqueKeys.push(
-        [...indexedColumns].sort((left, right) => Number(left.seqno) - Number(right.seqno)).map(column => column.name),
-      );
-    }
-    const compatibleUniqueKeys = (contract.uniqueKeys ?? []).every(expected =>
-      actualUniqueKeys.some(
-        actual => actual.length === expected.length && actual.every((column, index) => column === expected[index]),
-      ),
-    );
-    const compatibleDefinition = (contract.requiredDefinitionPatterns ?? []).every(pattern => pattern.test(definition));
-    const expectsWithoutRowid = contract.withoutRowid ?? true;
-    return {
-      compatible:
-        compatibleColumns &&
-        compatibleForeignKeys &&
-        compatibleUniqueKeys &&
-        compatibleDefinition &&
-        /\bWITHOUT\s+ROWID\b/i.test(definition) === expectsWithoutRowid,
-      exists: true,
-      group: contract.group,
-      name: contract.name,
-    };
-  });
-}
-
-const inspectPersistentExtensionTables = Effect.fn('codeGraph.inspectPersistentExtensionTables')(function* (
-  sql: SqlClient.SqlClient,
-) {
-  return yield* Effect.forEach(PERSISTENT_EXTENSION_TABLES, contract =>
-    persistentExtensionTableInspection(sql, contract),
-  );
-});
-
-export const codeGraphPersistentExtensionSchemaCompatible = Effect.fn('codeGraph.persistentExtensionSchemaCompatible')(
-  function* (sql: SqlClient.SqlClient) {
-    const inspections = yield* inspectPersistentExtensionTables(sql);
-    return inspections.every(inspection => inspection.exists && inspection.compatible);
-  },
-);
 
 const migratePersistentExtensionTables = Effect.fn('codeGraph.migratePersistentExtensionTables')(function* (
   sql: SqlClient.SqlClient,
@@ -3288,20 +2502,6 @@ const migratePersistentExtensionTables = Effect.fn('codeGraph.migratePersistentE
       }
     }),
   );
-});
-
-const validatePersistentExtensionTables = Effect.fn('codeGraph.validatePersistentExtensionTables')(function* (
-  sql: SqlClient.SqlClient,
-) {
-  const inspections = yield* inspectPersistentExtensionTables(sql);
-  const incompatible = inspections.filter(inspection => !inspection.exists || !inspection.compatible);
-  if (incompatible.length > 0) {
-    return yield* Effect.fail(
-      new CodeGraphStoreError(
-        `Code graph persistent extension schema is incompatible: ${incompatible.map(table => table.name).join(', ')}.`,
-      ),
-    );
-  }
 });
 
 const ensureSnapshotLeaseSchema = Effect.fn('codeGraph.ensureSnapshotLeaseSchema')(function* (
@@ -13263,39 +12463,6 @@ const ensureInitialReconciliationIndexes = Effect.fn('codeGraph.ensureInitialRec
     }
   }
 });
-
-function normalizeSchemaDefinition(value: string): string {
-  const quoted: string[] = [];
-  let unquoted = '';
-  for (let index = 0; index < value.length; index += 1) {
-    const opener = value[index]!;
-    const closer = opener === '[' ? ']' : opener;
-    if (opener !== "'" && opener !== '"' && opener !== '`' && opener !== '[') {
-      unquoted += opener;
-      continue;
-    }
-    const start = index;
-    for (index += 1; index < value.length; index += 1) {
-      if (value[index] !== closer) continue;
-      if (closer !== ']' && value[index + 1] === closer) {
-        index += 1;
-        continue;
-      }
-      break;
-    }
-    quoted.push(value.slice(start, Math.min(index + 1, value.length)));
-    unquoted += `\u0000${quoted.length - 1}\u0000`;
-  }
-  return unquoted
-    .toLowerCase()
-    .replace(/\bif not exists\b/gu, '')
-    .replace(/\s+/gu, ' ')
-    .replace(/\s*([(),])\s*/gu, '$1')
-    .trim()
-    .split('\u0000')
-    .map((segment, index) => (index % 2 === 1 ? (quoted[Number(segment)] ?? '') : segment))
-    .join('');
-}
 
 interface RemovedViewCleanupRow {
   readonly attempts: unknown;
